Testing a tutorial
It was working up til the point where I need to connect the turtlebot.

But still, haven't tried the option of launching the virtual turtlebot and see if it then was connected.
But I will put this tutorial to rest for now.

No need to merge this branch, leave it hanging until later.